+++ doc/release-notes.md
@@ -3,3 +3,4 @@ Bitcoin ABC version 0.18.8 is now available from:
   <https://download.bitcoinabc.org/0.18.8/>
 
 This release includes the following features and fixes:
+ - `dumpwallet` now includes hex-encoded scripts from the wallet in the dumpfile